requirements.txt
ffcv==1.0.2
matplotlib==3.7.2
mlflow==2.5.0
numpy==1.23.5
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.3.0
seaborn==0.13.2
skimage==0.0
submitit==1.5.2
timm==0.3.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
